Use dmp for hopefully faster changes
Used prop drilling (unfortunately) to pass props to client
to get client data.

Patches are created from changes which updates the text.

Because this still requires refreshing the
whole state on each edit, unclear how much
this will help with the rubber-banding
because this doesn't implement merging the way yjs does.