Swapped gateway ports 70 and 80 to prep for SSL
To run certbot on deployment, it requires the server to be running
on port 80. This change is to prep for certbot use on deployment.
Note that the change in port required the Dockerfile to expose 70:70
for the internal gateway again - as the final deployment uses kuber
this should not pose an issue.